pyproject.toml
[project]
name = "langgraphtest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.0.10",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
]
